requirements.txt
# ===========================================
# SuperSuite Production Requirements
# ===========================================
# This file contains all dependencies needed for the SuperSuite platform
# to run in production environments.

# ESSENTIAL DEPENDENCIES (Core Functionality)
# ===========================================

# Core dependencies
sqlmodel==0.0.14
pydantic==2.5.0
python-dotenv==1.0.0

# Database
snowflake-sqlalchemy==1.5.1
snowflake-connector-python==3.6.0
neo4j==5.18.0
psycopg2-binary==2.9.10

# AI/ML (Essential for LLM and embeddings)
transformers[agents]==4.40.0
sentence-transformers==2.7.0
openai==1.35.0
protobuf==3.20.3
torch>=2.0.0
tiktoken==0.8.0
huggingface-hub==0.27.1

# PDF Processing
PyPDF2==3.0.1

# Web UI (Essential for Streamlit app)
streamlit>=1.28.0

# Data Processing (Essential for dataframes and analysis)
numpy==1.24.3
pandas==2.0.3

# DEVELOPMENT & TESTING DEPENDENCIES
# ===================================

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.0

# Jupyter notebooks and UI
jupyter==1.0.0
ipykernel==6.27.1
ipywidgets==8.1.2

# CLI and formatting
rich==13.7.1
rich[jupyter]==13.7.1

# OPTIONAL DEPENDENCIES (Enhancements)
# ====================================

# Visualization
pyvis==0.3.2
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
networkx==3.1

# Streamlit enhancements
streamlit-extras>=0.3.0

# Utilities
python-dateutil==2.9.0
requests>=2.25.0
urllib3>=1.26.0

# Development tools
watchdog>=2.1.0
python-multipart>=0.0.5
